# docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Admin Platform API
  version: "1.0.0"
  description: |
    Backend API for the Admin Platform.
    All module endpoints require **Bearer JWT** authentication and pass through a
    four-layer guard chain: `authenticateJwt -> requireRole -> requireFeatureEnabled -> requirePermission`.

    **Error envelope** (all non-2xx responses):
    ```json
    { "code": "<ErrorCode>", "message": "<human string>" }
    ```

    **Pagination envelope** (list endpoints):
    ```json
    { "items": [...], "page": 1, "limit": 50, "total": 123 }
    ```
  contact:
    name: Admin Platform
servers:
  - url: http://localhost:7002
    description: Local development
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Two-tier JWT.  Tokens are signed with role-specific secrets
        (`JWT_SECRET_SUPER_ADMIN` / `JWT_SECRET_ADMIN`).
        Payload: `{ sub, email, role }`.

  # ─── Reusable schemas ───────────────────────────────────────────────
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }

    # ── Auth ──
...
    Task:
      type: object
      properties:
        _id: { type: string }
        title: { type: string }
        description: { type: string }
        projectId: { type: string }
        status: { type: string, enum: [todo, in_progress, review, done, cancelled] }
        priority: { type: string, enum: [low, medium, high, critical] }
        assigneeUserId: { type: string }
        reporterUserId: { type: string }
        dueDate: { type: string, format: date-time }
        completedAt: { type: string, format: date-time }
        tags: { type: array, items: { type: string } }
        estimatedHours: { type: number }
        createdAt: { type: string, format: date-time }

    # ── Calendar ──
...
  parameters:
    PageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    LimitParam:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: string }
      description: MongoDB ObjectId

  # ─── Reusable responses ───────────────────────────────────────────
...
paths:
...
  /api/v1/tasks/tasks:
    get:
      tags: [Tasks]
      summary: List tasks (paginated, filterable)
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - { name: status, in: query, schema: { type: string, enum: [todo, in_progress, review, done, cancelled] } }
        - { name: projectId, in: query, schema: { type: string } }
        - { name: assigneeUserId, in: query, schema: { type: string } }
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginationMeta"
                  - type: object
                    properties:
                      items: { type: array, items: { $ref: "#/components/schemas/Task" } }
    post:
      tags: [Tasks]
      summary: Create task (reporterUserId auto-set)
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
  /api/v1/tasks/tasks/{id}:
    parameters: [{ $ref: "#/components/parameters/IdPath" }]
    get:
      tags: [Tasks]
      summary: Get task
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
    patch:
      tags: [Tasks]
      summary: Update task
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
    delete:
      tags: [Tasks]
      summary: Delete task (done/cancelled only)
      responses:
        "204": { description: Deleted }
  /api/v1/tasks/tasks/{id}/transition:
    parameters: [{ $ref: "#/components/parameters/IdPath" }]
    post:
      tags: [Tasks]
      summary: Transition task status
      description: |
        todo->[in_progress,cancelled], in_progress->[review,todo,cancelled],
        review->[done,in_progress,cancelled], done->[todo], cancelled->[todo].
        Sets/clears completedAt automatically.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [to]
              properties:
                to: { type: string, enum: [todo, in_progress, review, done, cancelled] }
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
...